CMPE-275-Project/requirements.txt
Flask
requests
cassandra-driver
mmh3
ConfigParser